requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
chromadb==0.5.7
rank-bm25==0.2.2
openai==1.51.0
python-dotenv==1.0.1
PyMuPDF==1.24.10
streamlit==1.38.0
requests==2.32.3
